LU-3285 llite: check layout size after cl_object_layout_get
Check that layout size is not zero before doing flush for
DoM object. Also patch does that:
- initializes all values in cl_layout explicitly
  in lov_object_layout_get().
- changes cl_dom_comp_size to u64 for future needs, e.g.
  importing an existing filesystem image directly.

Test-Parameters: mdssizegb=20 testlist=sanity-dom,dom-performance
Reviewed-on: https://review.whamcloud.com/29810
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/include/cl_object.h
@@ -290,7 +290,7 @@ struct cl_layout {
 	/** size of layout in lov_mds_md format. */
 	size_t		cl_size;
 	/** size of DoM component if exists or zero otherwise */
-	u32		cl_dom_comp_size;
+	u64		cl_dom_comp_size;
 	/** Layout generation. */
 	u32		cl_layout_gen;
 	/** whether layout is a composite one */
